Begin major refactoring of render pipeline
- Added render queue
- Added Renderable class, which encapsulates anything that can be
rendered (Models, Particles, Lighting etc.) to all be in the same
render queue. Part of OpenGL abstraction
- Added Model class, which is currently defined as anything with a
VertexBuffer (anything that is an object in space). Part of game engine
- Added Cube class, a derived class of Model, which provides specific
vertices and indices to draw a cube on the screen. Will eventually use
an actual .obj file for that data. Part of sandbox, which is a folder
for testing parts of the game engine.
- Tried to use const wherever possible